Moves data classes into own files and some cleanup
