+++ Assets/WebVR/Scripts/WebVRCamera.cs
@@ -1,4 +1,5 @@
 ﻿using UnityEngine;
+using UnityEngine.XR;
 using System.Linq;
 using System.Collections;
 using System.Runtime.InteropServices;
