+++ Assets/WebVR/Scripts/WebVRControllerData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d5c7700e6bf9dfa40950b18dc3bddbcb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
