+++ Assets/WebVR/Scripts/WebVRData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3a7dd790407e5ef449d8629d7209c101
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
